requirements.txt
Django==3.1.4
djangorestframework==3.12.2
flake8==3.8.4
PyJWT==1.7.1